refactor: rename file agents/analyst -> agents/analyst_data
